# docs/site/public/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateDomainClaimRequest": {
        "properties": {
          "hostname": {
            "title": "Hostname",
            "type": "string"
          },
          "mode": {
            "anyOf": [
              {
                "enum": [
                  "direct",
                  "cloudflare"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mode"
          }
        },
        "required": [
          "hostname"
        ],
        "title": "CreateDomainClaimRequest",
        "type": "object"
      },
...
      "CustomDomainCapabilityResponse": {
        "properties": {
          "admission_enabled": {
            "title": "Admission Enabled",
            "type": "boolean"
          },
          "automatic": {
            "$ref": "#/components/schemas/AutomaticDomainTransitionCapability"
          },
          "cloudflare": {
            "$ref": "#/components/schemas/CloudflareCapability"
          },
          "control_ready": {
            "title": "Control Ready",
            "type": "boolean"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "routing_enabled": {
            "title": "Routing Enabled",
            "type": "boolean"
          },
          "routing_targets": {
            "items": {
              "$ref": "#/components/schemas/CustomDomainRoutingTarget"
            },
            "title": "Routing Targets",
            "type": "array"
          },
          "status": {
            "enum": [
              "disabled",
              "unready",
              "ready"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "detail",
          "enabled",
          "control_ready",
          "admission_enabled",
          "routing_enabled",
          "routing_targets",
          "automatic",
          "cloudflare"
        ],
        "title": "CustomDomainCapabilityResponse",
        "type": "object"
      },
      "CustomDomainRoutingTarget": {
        "properties": {
          "type": {
            "enum": [
              "A",
              "AAAA"
            ],
            "title": "Type",
            "type": "string"
          },
          "value": {
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "type",
          "value"
        ],
        "title": "CustomDomainRoutingTarget",
        "type": "object"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "A Buzz session or deployment token. Most operations require a session. Deployment tokens are accepted only for deployment to their assigned site.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HTTP API for deploying and managing sites on a self-hosted Buzz server. API operations are available only on the configured Buzz domain.",
    "title": "Buzz",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/capabilities/custom-domains": {
      "get": {
        "operationId": "getCustomDomainCapability",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomDomainCapabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Check custom-domain capability",
        "tags": [
          "Custom Domains"
        ]
      }
    },
...
    "/sites/{site_name}/domains": {
      "get": {
        "operationId": "listDomainClaims",
        "parameters": [
          {
            "in": "path",
            "name": "site_name",
            "required": true,
            "schema": {
              "title": "Site Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DomainClaimResponse"
                  },
                  "title": "Response Listdomainclaims",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List custom domain claims",
        "tags": [
          "Custom Domains"
        ]
      },
      "post": {
        "operationId": "createDomainClaim",
        "parameters": [
          {
            "in": "path",
            "name": "site_name",
            "required": true,
            "schema": {
              "title": "Site Name",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDomainClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainClaimResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too Many Requests"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create a custom domain claim",
        "tags": [
          "Custom Domains"
        ]
      }
    },
    "/sites/{site_name}/domains/{claim_id}": {
      "delete": {
        "operationId": "cancelDomainClaim",
        "parameters": [
          {
            "in": "path",
            "name": "site_name",
            "required": true,
            "schema": {
              "title": "Site Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Cancel a custom domain claim",
        "tags": [
          "Custom Domains"
        ]
      }
    },
    "/sites/{site_name}/domains/{claim_id}/check": {
      "post": {
        "operationId": "checkDomainClaim",
        "parameters": [
          {
            "in": "path",
            "name": "site_name",
            "required": true,
            "schema": {
              "title": "Site Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainClaimResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Check custom domain ownership",
        "tags": [
          "Custom Domains"
        ]
      }
    },
    "/sites/{site_name}/domains/{claim_id}/transition/cancel": {
      "post": {
        "operationId": "cancelDomainTransition",
        "parameters": [
          {
            "in": "path",
            "name": "site_name",
            "required": true,
            "schema": {
              "title": "Site Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainClaimResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Cancel a custom domain transition",
        "tags": [
          "Custom Domains"
        ]
      }
    },
    "/sites/{site_name}/domains/{claim_id}/transition/retry": {
      "post": {
        "operationId": "retryDomainTransition",
        "parameters": [
          {
            "in": "path",
            "name": "site_name",
            "required": true,
            "schema": {
              "title": "Site Name",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DomainClaimResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Retry a custom domain transition",
        "tags": [
          "Custom Domains"
        ]
      }
    },
...
  },
...
}